move down files from /provable
