--- src/provable/provable-bigint.ts
@@ -1,6 +1,0 @@
-import { createProvable } from './provable-generic.js';
-import { Field } from './field-bigint.js';
-
-export { provable };
-
-let provable = createProvable<Field>();
